requirements.txt
requests>=2.28.1
gitpython>=3.1.30
pathlib>=1.0.1
schedule>=1.1.0
loguru>=0.6.0
rich>=12.6.0
python-dotenv>=0.21.0
